+++ 2019/day_4_secure_container_1.cpp
@@ -19,10 +19,10 @@ next(string& current) {
 
 bool
 is_valid(const std::string& v) {
-    for(int i = 0; i < v.size() - 1; ++i)
+    for(size_t i = 0; i < v.size() - 1; ++i)
         if (v[i] > v[i+1]) return false;
     
-    for(int i = 0; i < v.size() - 1; ++i)
+    for(size_t i = 0; i < v.size() - 1; ++i)
         if (v[i] == v[i+1]) return true;
     
     return false;
